refactor(dilesa): venta detail — drawer → página completa con más espacio
- Nueva `app/dilesa/ventas/[id]/page.tsx` con las mismas 5 secciones
  (cliente, ficha, pipeline, pagos, expediente) en layout `max-w-6xl`
  y grids de 3 columnas; sigue el patrón de
  `app/rdb/inventario/levantamientos/[id]/page.tsx`.
- Click en una fila de `VentasModule` ahora navega vía
  `router.push('/dilesa/ventas/<id>')` en lugar de abrir un drawer.
- Borrado `venta-detail-drawer.tsx` (sin más referencias).
- Cleanup: removidos los `console.error/log` diagnósticos que se
  habían dejado mientras se cazaba el bug de URL > 8KB; el comentario
  inline en `ventas-module.tsx` ya conserva la razón del fix.
