+++ app/dilesa/ventas/[id]/page.tsx
@@ -1,0 +1,647 @@
+'use client';
+
+/* eslint-disable react-hooks/set-state-in-effect --
+ * Mismo data-sync pattern que el resto de páginas de detalle (cf.
+ * app/rdb/inventario/levantamientos/[id]/page.tsx).
+ */
+
+/**
+ * Detalle completo de una venta DILESA — 5 secciones:
+ *   1. Datos del cliente (`erp.personas`, cross-schema).
+ *   2. Datos de la venta — ficha + KYC/PLD + notas.
+ *   3. Pipeline — `<ActivityLog>` alimentado por `dilesa.venta_fases`.
+ *   4. Pagos — `dilesa.venta_pagos` con sus adjuntos.
+ *   5. Expediente digital — `erp.adjuntos` agrupados por rol.
+ *
+ * Lectura pura — captura/edición es entregable posterior.
+ */
+
+import Link from 'next/link';
+import { useParams } from 'next/navigation';
+import { useEffect, useMemo, useState } from 'react';
+import { ArrowLeft, ExternalLink, FileText } from 'lucide-react';
+import { RequireAccess } from '@/components/require-access';
+import { createSupabaseBrowserClient } from '@/lib/supabase-browser';
+import { ActivityLog } from '@/components/activity-log/activity-log';
+import type { ActivityEvent } from '@/components/activity-log/types';
+import { Badge } from '@/components/ui/badge';
+import type { BadgeTone } from '@/components/ui/badge';
+import { Skeleton } from '@/components/ui/skeleton';
+import { getAdjuntoProxyUrl } from '@/lib/adjuntos';
+import { getSupabaseErrorMessage } from '@/lib/supabase-error';
+
+type Venta = {
+  id: string;
+  persona_id: string;
+  unidad_id: string | null;
+  estado: string;
+  fase_actual: string | null;
+  fase_posicion: number | null;
+  tipo_credito: string | null;
+  valor_comercial: number | null;
+  valor_escrituracion: number | null;
+  precio_asignacion: number | null;
+  monto_credito_titular: number | null;
+  monto_credito_cotitular: number | null;
+  credito_titular_ref: string | null;
+  credito_cotitular_ref: string | null;
+  enganche_requerido: number | null;
+  descuento_total: number | null;
+  comision_vendedor: number | null;
+  comision_gerencia: number | null;
+  anticipo_comision: number | null;
+  monto_avaluo: number | null;
+  gastos_escrituracion: number | null;
+  numero_escritura: string | null;
+  fecha_escritura: string | null;
+  vendedor: string | null;
+  notario: string | null;
+  casa_valuadora: string | null;
+  es_pep: boolean | null;
+  ocupacion: string | null;
+  ine_numero: string | null;
+  forma_pago: string | null;
+  uso_efectivo: string | null;
+  conocimiento_dueno_beneficiario: string | null;
+  motivo_desasignacion: string | null;
+  notas: string | null;
+};
+
+type Persona = {
+  nombre: string | null;
+  apellido_paterno: string | null;
+  apellido_materno: string | null;
+  email: string | null;
+  telefono: string | null;
+  curp: string | null;
+  rfc: string | null;
+  nss: string | null;
+  fecha_nacimiento: string | null;
+  nacionalidad: string | null;
+  tipo_persona: string | null;
+  estado_civil: string | null;
+  domicilio: string | null;
+};
+
+type UnidadInfo = { identificador: string; proyecto_id: string | null };
+type Fase = { id: string; fase: string; posicion: number | null; fecha: string | null };
+type Pago = { id: string; fecha: string | null; monto: number; tipo: string | null };
+type Adjunto = {
+  id: string;
+  entidad_tipo: string;
+  entidad_id: string;
+  rol: string;
+  nombre: string;
+  url: string;
+  tipo_mime: string | null;
+};
+
+const ESTADO_TONE: Record<string, BadgeTone> = {
+  activa: 'info',
+  desasignada: 'neutral',
+};
+const ESTADO_LABEL: Record<string, string> = {
+  activa: 'Activa',
+  desasignada: 'Desasignada',
+};
+
+const ROL_LABEL: Record<string, string> = {
+  factura: 'Factura',
+  aprobacion_credito: 'Aprobación de crédito',
+  constancia_credito_titular: 'Constancia de crédito (titular)',
+  constancia_credito_cotitular: 'Constancia de crédito (co-titular)',
+  aviso_pld: 'Aviso PLD',
+  avaluo_comercial: 'Avalúo comercial',
+  contrato_promesa: 'Contrato promesa de compraventa',
+  solicitud_asignacion: 'Solicitud de asignación',
+  recibos_caja: 'Recibos de caja',
+  expediente_digital: 'Expediente digital',
+  ficu: 'FICU',
+  aviso_privacidad: 'Aviso de privacidad',
+  carta_instruccion_notarial: 'Carta instrucción notarial',
+  checklist_entrega: 'Checklist de entrega',
+  checklist_pre_entrega: 'Checklist pre-entrega',
+  validacion_patronal: 'Validación patronal',
+  nota_credito: 'Nota de crédito',
+  pagare: 'Pagaré',
+  imagen_detonacion: 'Imagen de detonación',
+  recibo_caja: 'Recibo de caja',
+  comprobante_deposito: 'Comprobante de depósito',
+};
+
+const moneyFmt = new Intl.NumberFormat('es-MX', {
+  style: 'currency',
+  currency: 'MXN',
+  maximumFractionDigits: 0,
+});
+
+const PIPELINE_TONES = {
+  fase_pipeline: { label: 'Pipeline', tone: 'info' as BadgeTone },
+};
+
+function fmtMoney(n: number | null): string | null {
+  return n == null ? null : moneyFmt.format(n);
+}
+
+function fmtFecha(s: string | null): string | null {
+  if (!s) return null;
+  const d = new Date(`${s}T00:00:00`);
+  if (isNaN(d.getTime())) return s;
+  return d.toLocaleDateString('es-MX', { day: '2-digit', month: 'short', year: 'numeric' });
+}
+
+/**
+ * @module Venta detail (DILESA)
+ * @responsive desktop-only
+ */
+export default function VentaDetailPage() {
+  return (
+    <RequireAccess empresa="dilesa" modulo="dilesa.ventas">
+      <DetailInner />
+    </RequireAccess>
+  );
+}
+
+function DetailInner() {
+  const params = useParams<{ id: string }>();
+  const id = params.id;
+
+  const [venta, setVenta] = useState<Venta | null>(null);
+  const [persona, setPersona] = useState<Persona | null>(null);
+  const [unidad, setUnidad] = useState<UnidadInfo | null>(null);
+  const [proyectoNombre, setProyectoNombre] = useState<string | null>(null);
+  const [fases, setFases] = useState<Fase[]>([]);
+  const [pagos, setPagos] = useState<Pago[]>([]);
+  const [adjuntos, setAdjuntos] = useState<Adjunto[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    if (!id) return;
+    let activo = true;
+    const sb = createSupabaseBrowserClient();
+    setLoading(true);
+    setError(null);
+
+    (async () => {
+      const { data: vRow, error: vErr } = await sb
+        .schema('dilesa')
+        .from('ventas')
+        .select('*')
+        .eq('id', id)
+        .is('deleted_at', null)
+        .maybeSingle();
+      if (!activo) return;
+      if (vErr) {
+        setError(getSupabaseErrorMessage(vErr, 'No se pudo cargar la venta.'));
+        setLoading(false);
+        return;
+      }
+      if (!vRow) {
+        setError('Venta no encontrada.');
+        setLoading(false);
+        return;
+      }
+      const ventaRow = vRow as unknown as Venta;
+      setVenta(ventaRow);
+
+      const [pRes, fRes, pagosRes, uRes] = await Promise.all([
+        sb
+          .schema('erp')
+          .from('personas')
+          .select(
+            'nombre, apellido_paterno, apellido_materno, email, telefono, curp, rfc, nss, fecha_nacimiento, nacionalidad, tipo_persona, estado_civil, domicilio'
+          )
+          .eq('id', ventaRow.persona_id)
+          .maybeSingle(),
+        sb
+          .schema('dilesa')
+          .from('venta_fases')
+          .select('id, fase, posicion, fecha')
+          .eq('venta_id', ventaRow.id)
+          .is('deleted_at', null)
+          .order('posicion', { ascending: true }),
+        sb
+          .schema('dilesa')
+          .from('venta_pagos')
+          .select('id, fecha, monto, tipo')
+          .eq('venta_id', ventaRow.id)
+          .is('deleted_at', null)
+          .order('fecha', { ascending: true }),
+        ventaRow.unidad_id
+          ? sb
+              .schema('dilesa')
+              .from('unidades')
+              .select('identificador, proyecto_id')
+              .eq('id', ventaRow.unidad_id)
+              .maybeSingle()
+          : Promise.resolve({ data: null, error: null }),
+      ]);
+      if (!activo) return;
+
+      const firstErr = pRes.error ?? fRes.error ?? pagosRes.error ?? uRes.error;
+      if (firstErr) {
+        setError(getSupabaseErrorMessage(firstErr, 'No se pudo cargar el detalle de la venta.'));
+        setLoading(false);
+        return;
+      }
+
+      setPersona((pRes.data as unknown as Persona) ?? null);
+      setFases((fRes.data ?? []) as Fase[]);
+      setPagos((pagosRes.data ?? []) as Pago[]);
+      const uData = uRes.data as { identificador: string; proyecto_id: string | null } | null;
+      setUnidad(uData);
+
+      if (uData?.proyecto_id) {
+        const { data: prj } = await sb
+          .schema('dilesa')
+          .from('proyectos')
+          .select('nombre')
+          .eq('id', uData.proyecto_id)
+          .maybeSingle();
+        if (!activo) return;
+        setProyectoNombre((prj?.nombre as string | null) ?? null);
+      } else {
+        setProyectoNombre(null);
+      }
+
+      const pagoIds = ((pagosRes.data ?? []) as Pago[]).map((p) => p.id);
+      const allIds = [ventaRow.id, ...pagoIds];
+      const { data: adjRows, error: adjErr } = await sb
+        .schema('erp')
+        .from('adjuntos')
+        .select('id, entidad_tipo, entidad_id, rol, nombre, url, tipo_mime')
+        .in('entidad_tipo', ['venta', 'venta_pago'])
+        .in('entidad_id', allIds);
+      if (!activo) return;
+      if (adjErr) {
+        setError(getSupabaseErrorMessage(adjErr, 'No se pudieron cargar los adjuntos.'));
+        setLoading(false);
+        return;
+      }
+      setAdjuntos((adjRows ?? []) as Adjunto[]);
+      setLoading(false);
+    })();
+
+    return () => {
+      activo = false;
+    };
+  }, [id]);
+
+  const clienteNombre = useMemo(() => {
+    if (!persona) return '';
+    return (
+      [persona.nombre, persona.apellido_paterno, persona.apellido_materno]
+        .filter(Boolean)
+        .join(' ') || '(sin nombre)'
+    );
+  }, [persona]);
+
+  const pipelineEvents = useMemo<ActivityEvent[]>(
+    () =>
+      fases
+        .filter((f) => f.fecha)
+        .sort((a, b) => (a.posicion ?? 0) - (b.posicion ?? 0))
+        .map((f) => ({
+          id: f.id,
+          at: new Date(`${f.fecha}T12:00:00`).toISOString(),
+          type: 'fase_pipeline',
+          actor: null,
+          summary: `${f.posicion ? `${f.posicion}. ` : ''}${f.fase}`,
+        })),
+    [fases]
+  );
+
+  const adjuntosVenta = useMemo(
+    () => adjuntos.filter((a) => a.entidad_tipo === 'venta'),
+    [adjuntos]
+  );
+  const adjuntosPorRol = useMemo(() => {
+    const m = new Map<string, Adjunto[]>();
+    for (const a of adjuntosVenta) {
+      const arr = m.get(a.rol) ?? [];
+      arr.push(a);
+      m.set(a.rol, arr);
+    }
+    return [...m.entries()].sort((a, b) =>
+      (ROL_LABEL[a[0]] ?? a[0]).localeCompare(ROL_LABEL[b[0]] ?? b[0])
+    );
+  }, [adjuntosVenta]);
+
+  const adjuntosPorPago = useMemo(() => {
+    const m = new Map<string, Adjunto[]>();
+    for (const a of adjuntos.filter((x) => x.entidad_tipo === 'venta_pago')) {
+      const arr = m.get(a.entidad_id) ?? [];
+      arr.push(a);
+      m.set(a.entidad_id, arr);
+    }
+    return m;
+  }, [adjuntos]);
+
+  const totalPagos = useMemo(() => pagos.reduce((s, p) => s + (p.monto ?? 0), 0), [pagos]);
+
+  if (loading) {
+    return (
+      <div className="container mx-auto max-w-6xl space-y-6 px-4 py-6">
+        <Skeleton className="h-5 w-40" />
+        <Skeleton className="h-24 w-full rounded-lg" />
+        <Skeleton className="h-64 w-full rounded-lg" />
+        <Skeleton className="h-48 w-full rounded-lg" />
+      </div>
+    );
+  }
+
+  if (error || !venta) {
+    return (
+      <div className="container mx-auto max-w-6xl space-y-4 px-4 py-6">
+        <BackLink />
+        <div className="rounded-lg border border-destructive/40 bg-destructive/5 p-4 text-sm text-destructive">
+          {error ?? 'Venta no encontrada.'}
+        </div>
+      </div>
+    );
+  }
+
+  const fichaVenta: { label: string; value: string }[] = (
+    [
+      ['Proyecto', proyectoNombre],
+      ['Unidad', unidad?.identificador ?? null],
+      ['Tipo de crédito', venta.tipo_credito],
+      ['Vendedor', venta.vendedor],
+      ['Notario', venta.notario],
+      ['Casa valuadora', venta.casa_valuadora],
+      ['Precio de asignación', fmtMoney(venta.precio_asignacion)],
+      ['Valor comercial', fmtMoney(venta.valor_comercial)],
+      ['Valor de escrituración', fmtMoney(venta.valor_escrituracion)],
+      ['Enganche requerido', fmtMoney(venta.enganche_requerido)],
+      ['Descuento total', fmtMoney(venta.descuento_total)],
+      ['Crédito titular', fmtMoney(venta.monto_credito_titular)],
+      ['Crédito co-titular', fmtMoney(venta.monto_credito_cotitular)],
+      ['Ref. crédito titular', venta.credito_titular_ref],
+      ['Ref. crédito co-titular', venta.credito_cotitular_ref],
+      ['Comisión vendedor', fmtMoney(venta.comision_vendedor)],
+      ['Comisión gerencia', fmtMoney(venta.comision_gerencia)],
+      ['Anticipo comisión', fmtMoney(venta.anticipo_comision)],
+      ['Monto avalúo', fmtMoney(venta.monto_avaluo)],
+      ['Gastos escrituración', fmtMoney(venta.gastos_escrituracion)],
+      ['# Escritura', venta.numero_escritura],
+      ['Fecha de escritura', fmtFecha(venta.fecha_escritura)],
+    ] as [string, string | null][]
+  )
+    .filter((r): r is [string, string] => r[1] != null && r[1] !== '')
+    .map(([label, value]) => ({ label, value }));
+
+  const fichaPersona: { label: string; value: string }[] = persona
+    ? (
+        [
+          ['CURP', persona.curp],
+          ['RFC', persona.rfc],
+          ['NSS', persona.nss],
+          ['Tel.', persona.telefono],
+          ['Email', persona.email],
+          ['Fecha de nacimiento', fmtFecha(persona.fecha_nacimiento)],
+          ['Nacionalidad', persona.nacionalidad],
+          ['Estado civil', persona.estado_civil],
+          ['Tipo persona', persona.tipo_persona],
+          ['Domicilio', persona.domicilio],
+        ] as [string, string | null][]
+      )
+        .filter((r): r is [string, string] => r[1] != null && r[1] !== '')
+        .map(([label, value]) => ({ label, value }))
+    : [];
+
+  const kyc: { label: string; value: string }[] = (
+    [
+      ['PEP', venta.es_pep == null ? null : venta.es_pep ? 'Sí' : 'No'],
+      ['Ocupación', venta.ocupacion],
+      ['INE', venta.ine_numero],
+      ['Forma de pago', venta.forma_pago],
+      ['Uso de efectivo', venta.uso_efectivo],
+      ['Dueño beneficiario', venta.conocimiento_dueno_beneficiario],
+    ] as [string, string | null][]
+  )
+    .filter((r): r is [string, string] => r[1] != null && r[1] !== '')
+    .map(([label, value]) => ({ label, value }));
+
+  return (
+    <div className="container mx-auto max-w-6xl space-y-6 px-4 py-6">
+      <BackLink />
+
+      <header className="flex flex-wrap items-start justify-between gap-3">
+        <div>
+          <h1 className="text-2xl font-semibold tracking-tight text-[var(--text)]">
+            {clienteNombre || '(sin nombre)'}
+          </h1>
+          {proyectoNombre && unidad?.identificador ? (
+            <p className="mt-1 text-sm text-[var(--text)]/60">
+              {proyectoNombre} · {unidad.identificador}
+            </p>
+          ) : null}
+        </div>
+        <div className="flex flex-wrap items-center gap-1.5">
+          {venta.fase_actual ? (
+            <Badge tone="neutral">
+              {venta.fase_posicion ? `${venta.fase_posicion}. ` : ''}
+              {venta.fase_actual}
+            </Badge>
+          ) : null}
+          <Badge tone={ESTADO_TONE[venta.estado] ?? 'neutral'}>
+            {ESTADO_LABEL[venta.estado] ?? venta.estado}
+          </Badge>
+          {venta.tipo_credito ? <Badge tone="neutral">{venta.tipo_credito}</Badge> : null}
+        </div>
+      </header>
+
+      <Section title="Datos del cliente">
+        {fichaPersona.length === 0 ? (
+          <p className="text-sm text-[var(--text)]/60">Sin datos del cliente.</p>
+        ) : (
+          <FichaGrid rows={fichaPersona} cols={3} />
+        )}
+      </Section>
+
+      <Section title="Datos de la venta">
+        {fichaVenta.length === 0 ? (
+          <p className="text-sm text-[var(--text)]/60">—</p>
+        ) : (
+          <FichaGrid rows={fichaVenta} cols={3} />
+        )}
+        {venta.motivo_desasignacion ? (
+          <div className="mt-4 border-t border-[var(--border)] pt-4">
+            <div className="text-xs font-medium uppercase tracking-wide text-[var(--text)]/50">
+              Motivo de desasignación
+            </div>
+            <p className="mt-0.5 text-sm text-[var(--text)]/80">{venta.motivo_desasignacion}</p>
+          </div>
+        ) : null}
+        {kyc.length > 0 ? (
+          <div className="mt-6 border-t border-[var(--border)] pt-4">
+            <div className="mb-2 text-xs font-medium uppercase tracking-wide text-[var(--text)]/50">
+              KYC / PLD
+            </div>
+            <FichaGrid rows={kyc} cols={3} />
+          </div>
+        ) : null}
+        {venta.notas ? (
+          <div className="mt-4 border-t border-[var(--border)] pt-4">
+            <div className="text-xs font-medium uppercase tracking-wide text-[var(--text)]/50">
+              Notas
+            </div>
+            <p className="mt-0.5 whitespace-pre-wrap text-sm text-[var(--text)]/80">
+              {venta.notas}
+            </p>
+          </div>
+        ) : null}
+      </Section>
+
+      <Section title="Pipeline" description={`${pipelineEvents.length} de 17 fases alcanzadas`}>
+        <ActivityLog events={pipelineEvents} tones={PIPELINE_TONES} />
+      </Section>
+
+      <Section
+        title="Pagos"
+        description={
+          pagos.length === 0 ? 'sin pagos' : `${pagos.length} · ${moneyFmt.format(totalPagos)}`
+        }
+      >
+        {pagos.length === 0 ? (
+          <p className="text-sm text-[var(--text)]/60">No hay depósitos registrados.</p>
+        ) : (
+          <table className="w-full text-sm">
+            <thead>
+              <tr className="border-b border-[var(--border)] text-left text-xs uppercase tracking-wide text-[var(--text)]/50">
+                <th className="py-1 pr-2 font-medium">Fecha</th>
+                <th className="py-1 pr-2 font-medium">Tipo</th>
+                <th className="py-1 text-right font-medium">Monto</th>
+                <th className="py-1 pl-2 font-medium">Adjuntos</th>
+              </tr>
+            </thead>
+            <tbody>
+              {pagos.map((p) => {
+                const ads = adjuntosPorPago.get(p.id) ?? [];
+                return (
+                  <tr key={p.id} className="border-b border-[var(--border)]/40">
+                    <td className="py-1.5 pr-2">{fmtFecha(p.fecha) ?? '—'}</td>
+                    <td className="py-1.5 pr-2 text-[var(--text)]/70">{p.tipo ?? '—'}</td>
+                    <td className="py-1.5 text-right tabular-nums">{moneyFmt.format(p.monto)}</td>
+                    <td className="py-1.5 pl-2">
+                      <div className="flex flex-wrap gap-1">
+                        {ads.map((a) => (
+                          <AdjuntoLink key={a.id} a={a} compact />
+                        ))}
+                        {ads.length === 0 ? <span className="text-[var(--text)]/30">—</span> : null}
+                      </div>
+                    </td>
+                  </tr>
+                );
+              })}
+            </tbody>
+          </table>
+        )}
+      </Section>
+
+      <Section
+        title="Expediente digital"
+        description={
+          adjuntosVenta.length === 0 ? 'sin documentos' : `${adjuntosVenta.length} documentos`
+        }
+      >
+        {adjuntosPorRol.length === 0 ? (
+          <p className="text-sm text-[var(--text)]/60">
+            Sin documentos en el expediente para esta venta.
+          </p>
+        ) : (
+          <div className="space-y-4">
+            {adjuntosPorRol.map(([rol, ads]) => (
+              <div key={rol}>
+                <div className="mb-1.5 text-xs font-medium uppercase tracking-wide text-[var(--text)]/50">
+                  {ROL_LABEL[rol] ?? rol}
+                </div>
+                <ul className="flex flex-wrap gap-2">
+                  {ads.map((a) => (
+                    <li key={a.id}>
+                      <AdjuntoLink a={a} />
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            ))}
+          </div>
+        )}
+      </Section>
+    </div>
+  );
+}
+
+function BackLink() {
+  return (
+    <Link
+      href="/dilesa/ventas"
+      className="inline-flex items-center gap-1.5 text-sm text-[var(--text)]/60 hover:text-[var(--text)]"
+    >
+      <ArrowLeft className="size-4" /> Volver a ventas
+    </Link>
+  );
+}
+
+function Section({
+  title,
+  description,
+  children,
+}: {
+  title: string;
+  description?: string;
+  children: React.ReactNode;
+}) {
+  return (
+    <section className="rounded-lg border border-[var(--border)] bg-[var(--card)] p-5">
+      <div className="mb-4 flex items-baseline justify-between gap-3">
+        <h2 className="text-sm font-medium uppercase tracking-wider text-[var(--text)]/60">
+          {title}
+        </h2>
+        {description ? <span className="text-xs text-[var(--text)]/50">{description}</span> : null}
+      </div>
+      {children}
+    </section>
+  );
+}
+
+function FichaGrid({ rows, cols = 2 }: { rows: { label: string; value: string }[]; cols?: 2 | 3 }) {
+  const gridCls =
+    cols === 3
+      ? 'grid grid-cols-1 gap-x-6 gap-y-3 sm:grid-cols-2 lg:grid-cols-3'
+      : 'grid grid-cols-1 gap-x-6 gap-y-3 sm:grid-cols-2';
+  return (
+    <dl className={gridCls}>
+      {rows.map((r) => (
+        <div key={r.label}>
+          <dt className="text-xs font-medium uppercase tracking-wide text-[var(--text)]/50">
+            {r.label}
+          </dt>
+          <dd className="mt-0.5 text-sm text-[var(--text)]">{r.value}</dd>
+        </div>
+      ))}
+    </dl>
+  );
+}
+
+function AdjuntoLink({ a, compact = false }: { a: Adjunto; compact?: boolean }) {
+  const href = getAdjuntoProxyUrl(a.url);
+  return (
+    <a
+      href={href}
+      target="_blank"
+      rel="noopener noreferrer"
+      className={
+        compact
+          ? 'inline-flex items-center gap-1 rounded border border-[var(--border)] px-1.5 py-0.5 text-[10px] text-[var(--text)]/70 hover:text-[var(--text)]'
+          : 'inline-flex items-center gap-1.5 rounded-md border border-[var(--border)] bg-[var(--card)] px-2.5 py-1 text-xs text-[var(--text)]/80 hover:text-[var(--text)]'
+      }
+      title={a.nombre}
+    >
+      <FileText className={compact ? 'h-3 w-3' : 'h-3.5 w-3.5'} />
+      <span className="max-w-[220px] truncate">{a.nombre}</span>
+      <ExternalLink className={compact ? 'h-2.5 w-2.5' : 'h-3 w-3'} />
+    </a>
+  );
+}
